Fix integer comparison with intValue()

+++ codefights/src/isTreeSymmetric.java
@@ -1,14 +1,3 @@
-//
-// Definition for binary tree:
-// class Tree<T> {
-//   Tree(T x) {
-//     value = x;
-//   }
-//   T value;
-//   Tree<T> left;
-//   Tree<T> right;
-// }
-
 boolean isTreeSymmetric(Tree<Integer> t) {
     if (t == null) return true;
     return symmetricTrees(t.left, t.right);
@@ -18,7 +7,7 @@ boolean symmetricTrees(Tree<Integer> l, Tree<Integer> r) {
     if (l == null && r == null) {
         return true;
     }
-    if (l == null || r == null || l.value != r.value) {
+    if (l == null || r == null || l.value.intValue() != r.value.intValue()) {
         return false;
     }
     if (symmetricTrees(l.left, r.right) == false) {
